+++ package.json
@@ -1,17 +1,21 @@
 {
   "name": "pb-text-format-to-json",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "Protobuf text format to JSON parser.",
-  "main": "src/pb-parser.js",
+  "main": "dist/pb-parser.js",
   "type": "module",
   "dependencies": {
     "colors": "^1.4.0",
     "jest": "^26.6.3",
     "parsimmon": "^1.16.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@types/parsimmon": "^1.10.6",
+    "typescript": "^4.9.4"
+  },
   "scripts": {
-    "test": "node --experimental-vm-modules node_modules/.bin/jest"
+    "test": "node --experimental-vm-modules node_modules/.bin/jest",
+    "build": "tsc --build"
   },
   "repository": {
     "type": "git",
@@ -23,5 +27,9 @@
     "url": "https://github.com/k-j-kim/pb-parser/issues"
   },
   "homepage": "https://github.com/k-j-kim/pb-parser#readme",
-  "keywords": ["protobuf", "textformat", "json"]
+  "keywords": [
+    "protobuf",
+    "textformat",
+    "json"
+  ]
 }
